move configs.rawConfigProvider to raw_config_service package
